refactor(server): dedicated search vector side effects handlers
